third_party/blink/web_tests/platform/mac/external/wpt/editing/run/italic_1-1000-expected.txt

This is a testharness.js-based test.
Found 157 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["italic",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-style:italic\\">foo</span></p><p> <span style=\\"font-style:italic\\"><span>bar</span></span> </p><p><span style=\\"font-style:italic\\">baz</span></p>" but got "<p><span style=\\"font-style:italic\\">foo</span></p><p> <span style=\\"font-style:italic\\"><span>bar</span> </span></p><p><span style=\\"font-style:italic\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><i>foo</i></p><p> <i><span>bar</span></i> </p><p><i>baz</i></p>" but got "<p><i>foo</i></p><p> <i><span>bar</span> </i></p><p><i>baz</i></p>"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "<b>foo[]bar</b>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "<b>foo[]bar</b>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "<i>foo[]bar</i>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "<i>foo[]bar</i>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[bar]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[bar]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[bar]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[bar]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandIndeterm("italic") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "{<p><p> <p>foo</p>}" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "{<p><p> <p>foo</p>}" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "{<p><p> <p>foo</p>}" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "{<p><p> <p>foo</p>}" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<address>[bar]</address>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<address>[bar]</address>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<address>[bar]</address>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<address>[bar]</address>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<cite>[bar]</cite>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<cite><span style=\\"font-style:normal\\">bar</span></cite>baz" but got "foo<cite style=\\"font-style:normal\\">bar</cite>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<cite>[bar]</cite>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<cite>[bar]</cite>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<cite>[bar]</cite>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<cite><span style=\\"font-style:normal\\">bar</span></cite>baz" but got "foo<cite style=\\"font-style:normal\\">bar</cite>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<cite>[bar]</cite>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<cite>[bar]</cite>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<dfn>[bar]</dfn>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<dfn><span style=\\"font-style:normal\\">bar</span></dfn>baz" but got "foo<dfn style=\\"font-style:normal\\">bar</dfn>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<dfn>[bar]</dfn>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<dfn>[bar]</dfn>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<dfn>[bar]</dfn>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<dfn><span style=\\"font-style:normal\\">bar</span></dfn>baz" but got "foo<dfn style=\\"font-style:normal\\">bar</dfn>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<dfn>[bar]</dfn>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<dfn>[bar]</dfn>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo<em>[bar]</em>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "foo<em>[bar]</em>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<var>[bar]</var>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<var><span style=\\"font-style:normal\\">bar</span></var>baz" but got "foo<var style=\\"font-style:normal\\">bar</var>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<var>[bar]</var>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<var>[bar]</var>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<var>[bar]</var>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<var><span style=\\"font-style:normal\\">bar</span></var>baz" but got "foo<var style=\\"font-style:normal\\">bar</var>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<var>[bar]</var>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<var>[bar]</var>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<address>bar</address>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<address>bar</address>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<address>bar</address>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<address>bar</address>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<cite>bar</cite>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<cite><span style=\\"font-style:normal\\">bar</span></cite>baz" but got "foo<cite style=\\"font-style:normal\\">bar</cite>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<cite>bar</cite>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<cite>bar</cite>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<cite>bar</cite>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<cite><span style=\\"font-style:normal\\">bar</span></cite>baz" but got "foo<cite style=\\"font-style:normal\\">bar</cite>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<cite>bar</cite>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<cite>bar</cite>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<dfn>bar</dfn>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<dfn><span style=\\"font-style:normal\\">bar</span></dfn>baz" but got "foo<dfn style=\\"font-style:normal\\">bar</dfn>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<dfn>bar</dfn>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<dfn>bar</dfn>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<dfn>bar</dfn>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<dfn><span style=\\"font-style:normal\\">bar</span></dfn>baz" but got "foo<dfn style=\\"font-style:normal\\">bar</dfn>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<dfn>bar</dfn>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<dfn>bar</dfn>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo{<em>bar</em>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "foo{<em>bar</em>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<var>bar</var>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<var><span style=\\"font-style:normal\\">bar</span></var>baz" but got "foo<var style=\\"font-style:normal\\">bar</var>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<var>bar</var>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<var>bar</var>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<var>bar</var>}baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<var><span style=\\"font-style:normal\\">bar</span></var>baz" but got "foo<var style=\\"font-style:normal\\">bar</var>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<var>bar</var>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<var>bar</var>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<address>b[a]r</address>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<address>b[a]r</address>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<address>b[a]r</address>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<address>b[a]r</address>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<cite>b[a]r</cite>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<cite>b[a]r</cite>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<cite>b[a]r</cite>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<cite>b[a]r</cite>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<dfn>b[a]r</dfn>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<dfn>b[a]r</dfn>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<dfn>b[a]r</dfn>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<dfn>b[a]r</dfn>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<em>b[a]r</em>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<em>b[a]r</em>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<em>b[a]r</em>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<em>b[a]r</em>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>b[a]r</i>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-style:italic\\">b</span>a<span style=\\"font-style:italic\\">r</span>baz" but got "foo<i>b</i>a<i>r</i>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>b[a]r</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>b[a]r</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<i>b[a]r</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<i>b[a]r</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<var>b[a]r</var>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<var>b[a]r</var>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<var>b[a]r</var>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<var>b[a]r</var>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<address>bar</address>b]az" queryCommandIndeterm("italic") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<address>bar</address>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<address>bar</address>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<address>bar</address>b]az" queryCommandIndeterm("italic") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<address>bar</address>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<address>bar</address>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<cite>bar</cite>b]az" queryCommandIndeterm("italic") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<cite>bar</cite>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<cite>bar</cite>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<cite>bar</cite>b]az" queryCommandIndeterm("italic") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<cite>bar</cite>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<cite>bar</cite>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
Harness: the test ran to completion.
